Address PR reviews - use the existing getTokenFromFile func to retrieve the token if provided - semantic issues
